+++ configurio/tests/conftest.py
@@ -8,6 +8,18 @@ from . import core
 core.change_path_to_testsdir()
 
 
+def pytest_configure(config):
+    custom_markers = (
+        "json: run only json tests",
+        "yaml: run only yaml tests"
+    )
+
+    for marker in custom_markers:
+        config.addinivalue_line(
+            "markers", marker
+        )
+
+
 @pytest.fixture(scope="class", autouse=True)
 def dir_control():
     create_directories(core.temp_dir_path)