+++ configurio/tests/test_yaml.py
@@ -1,27 +1,29 @@
-import json
-
 import pytest
+from ruamel.yaml import YAML
 
-from configurio import JSON_Configuration
+from configurio import YAML_Configuration
 
 from .core import BaseConfigTest, configuration_file_path,\
     default_configuration_dict, default_cfg_json_path
 
 
-@pytest.mark.json
+yaml = YAML()
+
+
+@pytest.mark.yaml
 class Test_Create_DefDict(BaseConfigTest):
     """
     Create configuration file on `configuration_file_path`
     and get default config values from `default_configuration_dict`
     """
     def setup(self):
-        self.config = JSON_Configuration(
+        self.config = YAML_Configuration(
             configuration_file_path,
             default_configuration_dict
         )
 
 
-@pytest.mark.json
+@pytest.mark.yaml
 class Test_Create_DefPath(BaseConfigTest):
     """
     Read existing file on `configuration_file_path`
@@ -29,9 +31,9 @@ class Test_Create_DefPath(BaseConfigTest):
     """
     def setup(self):
         with open(default_cfg_json_path, 'w') as f:
-            json.dump(default_configuration_dict, f)
+            yaml.dump(default_configuration_dict, f)
 
-        self.config = JSON_Configuration(
+        self.config = YAML_Configuration(
             configuration_file_path,
             default_cfg_json_path
         )